[fix] removed the use of routes
