+++ frontend/gostarkme-web/state/nFunds.ts
@@ -1,0 +1,5 @@
+import { atomWithReset } from "jotai/utils"
+
+export const clickedFundState = atomWithReset<
+  Number | null | undefined
+>(undefined)
